fix(phase-6): resolve all audit blockers, gaps and inconsistencies
- coverage.md: add Phase 6 95% gate section with src/simulation/ preflight
- ci.yml: simulation_tests routing check, src/simulation/ preflight, 60-min timeout
- hud-layout.md: add getDemandPressurePct [0,1] display-scaling + INVERSE SEMANTICS note
- notification-system.md: add deficit-streak warnings to CRITICAL queue
- service-coverage.md: add service_degradation_probability_per_tick probability spec
- undo-system.md: add paused-simulation sentinel spec (tick-based expiry, recompute-on-unpause)
- simulation_constants.h: add service_degradation_probability_per_tick = 0.5f constant
- simulation_types.h: change SimulationNotification.amount float→int (treasury values are integers)
- AitownTestHelpers.cmake: update simulation_tests comment to DISCOVERY_TIMEOUT 60
- phase-6.md: 31 spec fixes from prod-owner/tech-squad audit (FIX 20-31) + design squad fixes
  - D-CRITICAL2: GameOver test switch to NiceMock + targeted triggerStinger negative assertion
  - D-CRITICAL3: UndoSystem_PausedSimulation test contract (tick-based expiry)
  - D-HIGH2: service_degradation_probability_per_tick constant + spec reference
  - D-HIGH4: deficit-streak warnings in CRITICAL queue parenthetical
  - D-MEDIUM1: SimulationNotification.amount float→int
  - D-MEDIUM2: BondIssuance_PreventsFurtherForcedLoans_UntilDebtCapClears test

+++ src/interfaces/simulation_types.h
@@ -98,7 +98,7 @@ enum class NotificationType {
 // Dequeued via ICitySimulation::pollPendingNotification() (singular, FIFO, one per call).
 struct SimulationNotification {
     NotificationType type{NotificationType::ForcedLoanIssued};
-    float  amount{0.0f};         // loan principal for loan types; 0 for others
+    int    amount{0};            // loan principal for loan types; 0 for others (int: all treasury values are integers; avoids float precision loss for large principals)
     int    repaymentTicks{0};    // repayment period for loan types; 0 for others
     int    milestoneValue{0};    // population count for PopulationMilestone;
                                  // CityRatingTier as int for CityRatingTransition; 0 for others
